refactor(keyboard): split keycap appearance into style files
