+++ Apps/Keyty/Sources/Keyty/Features/Visualizers/Keyboard/Appearance/KeycapAppearance+Retro.swift
@@ -1,0 +1,85 @@
+//
+//  KeycapAppearance+Retro.swift
+//  Keyty
+//
+//  SPDX-FileCopyrightText: 2026 Serhii Bykov
+//  SPDX-License-Identifier: BSD-3-Clause
+//
+
+import AppKit
+
+extension KeycapAppearance {
+    struct Retro {
+        let textColor: NSColor
+        let groupBackgroundColor: NSColor
+        let groupStrokeColor: NSColor
+        let badgeFillColor: NSColor
+        let badgeStrokeColor: NSColor
+        let badgeHighlightColor: NSColor
+        let badgeTextColor: NSColor
+        let lipGradient: NSGradient?
+        let bodyShadowColor: NSColor
+        let bodyGradient: NSGradient?
+        let bodyStrokeColor: NSColor
+        let faceShadowGradient: NSGradient?
+        let darkFaceGradient: NSGradient?
+        let darkFaceStrokeColor: NSColor
+        let lightFaceGradient: NSGradient?
+        let lightFaceStrokeColor: NSColor
+        let glossColor: NSColor
+
+        init(tokens: KeycapThemeTokens) {
+            let bodyEdge = tokens.surfaceShadowColor.darkened(by: 0.08)
+            let bodyCenter = tokens.surfaceBaseColor.darkened(by: 0.10)
+            let darkFaceTop = tokens.surfaceBaseColor.lightened(by: 0.03)
+            let darkFaceBottom = tokens.surfaceBaseColor.darkened(by: 0.01)
+            let lightFaceTop = tokens.surfaceHighlightColor.lightened(by: 0.08)
+            let lightFaceBottom = tokens.surfaceBaseColor
+            let groupBackgroundColor = tokens.recessColor.darkened(by: 0.40)
+            let groupStrokeColor = tokens.surfaceBaseColor.withAlphaComponent(0.70)
+            let lipEdge = tokens.recessColor.darkened(by: 0.34).withAlphaComponent(0.96)
+            let lipCenter = tokens.recessColor.darkened(by: 0.12).withAlphaComponent(0.86)
+            let bodyStrokeColor = tokens.surfaceBorderColor.withAlphaComponent(0.20)
+            let darkFaceStrokeColor = tokens.surfaceBorderColor.withAlphaComponent(0.20)
+            let faceShadowEdge = tokens.recessColor.darkened(by: 0.34).withAlphaComponent(0.46)
+            let faceShadowCenter = tokens.recessColor.darkened(by: 0.12).withAlphaComponent(0.26)
+
+            self.textColor = tokens.textColor
+            self.groupBackgroundColor = groupBackgroundColor
+            self.groupStrokeColor = groupStrokeColor
+            self.badgeFillColor = tokens.badgeFillColor
+            self.badgeStrokeColor = tokens.badgeStrokeColor
+            self.badgeHighlightColor = tokens.badgeHighlightColor
+            self.badgeTextColor = tokens.badgeTextColor
+            self.lipGradient = NSGradient(colorsAndLocations:
+                (lipEdge, 0.0),
+                (lipCenter, 0.5),
+                (lipEdge, 1.0)
+            )
+            self.bodyShadowColor = NSColor.black.withAlphaComponent(0.6)
+            self.bodyGradient = NSGradient(colorsAndLocations:
+                (bodyEdge, 0.0),
+                (bodyCenter, 0.32),
+                (bodyCenter, 0.68),
+                (bodyEdge, 1.0)
+            )
+            self.bodyStrokeColor = bodyStrokeColor
+            self.faceShadowGradient = NSGradient(colorsAndLocations:
+                (faceShadowEdge, 0.0),
+                (faceShadowCenter, 0.5),
+                (faceShadowEdge, 1.0)
+            )
+            self.darkFaceGradient = NSGradient(colorsAndLocations:
+                (darkFaceTop, 0.0),
+                (darkFaceBottom, 1.0)
+            )
+            self.darkFaceStrokeColor = darkFaceStrokeColor
+            self.lightFaceGradient = NSGradient(colorsAndLocations:
+                (lightFaceTop, 0.0),
+                (lightFaceBottom, 1.0)
+            )
+            self.lightFaceStrokeColor = tokens.surfaceBorderColor.withAlphaComponent(0.55)
+            self.glossColor = NSColor(white: 1, alpha: 0.08)
+        }
+    }
+}
